Fixed link script, added vim plugins

+++ link_them.sh
@@ -2,6 +2,9 @@
 
 TARGET=$HOME
 
-for f in *rc ; do
-    ln -sf "$(readlink -f $f)" "$TARGET/.$f";
+for f in * ; do
+    if [ "$(readlink -f $0)" != "$(readlink -f $f)" ]
+    then
+        ln -sf "$(readlink -f $f)" "$TARGET/.$f";
+    fi
 done
